+++ quartz-js/components/sidebar/docs/code/sidebar.js
@@ -1,3 +1,13 @@
 // Example Sidebar Component
 
-m.component(vhxm.components.shared.sidebar.ui.container);
+m.component(vhxm.components.shared.sidebar.ui.container, {
+  loadedState: vhxm.models.customers().single,
+  template: vhxm.customers.sidebar.ui.container,
+  onClose: function() {
+    m.route('/admin/customers');
+  },
+  onInit: function() {
+    vhxm.components.shared.sidebar.toggle('open');
+    vhxm.models.retrieveCustomer(m.route.param('id'));
+  }
+});