+++ quartz-js/components/sidebar/sidebar.controller.js
@@ -1,42 +1,49 @@
-vhxm.components.shared.sidebar.controller = function() {
-  var self = this;
+vhxm.components.shared.sidebar.controller = function(opts) {
+  let self = this;
+
+  if (opts) {
+    if (opts.onInit) {
+      opts.onInit();
+    }
+    if (opts.loadedState) {
+      vhxm.components.shared.sidebar.setLoadedState(opts.loadedState);
+    }
+    if (opts.template) {
+      vhxm.components.shared.sidebar.setTemplate(opts.template);
+    }
+    if (opts.onOpen) {
+      vhxm.components.shared.sidebar.state.onOpen(opts.onOpen);
+    }
+    if (opts.onClose) {
+      vhxm.components.shared.sidebar.state.onClose(opts.onClose);
+    }
+  }
 
   self.animatorIn = function(elem, isInit) {
-    // if (!isInit) {
-      if (vhxm.components.shared.sidebar.state.isOpen()) {
-        $(elem).velocity({
-          right: 0,
-        }, {
-          duration: 500,
-          easing: [ 0.19, 1, 0.22, 1 ],
-          complete: function() {
-            if (vhxm.components.shared.sidebar.state.animate_in_cb) {
-              vhxm.components.shared.sidebar.state.onAnimateIn();
-            }
-          }
-        });
-      }
-    // }
+    if (vhxm.components.shared.sidebar.state.isOpen()) {
+      $(elem).velocity({
+        right: 0
+      }, {
+        duration: vhxm.components.shared.sidebar.state.skipTransition() ? 0 : 500,
+        easing: [0.19, 1, 0.22, 1],
+        complete: function complete() {
+          vhxm.components.shared.sidebar.state.onOpen();
+          vhxm.components.shared.sidebar.state.skipTransition(true);
+        }
+      });
+    }
   };
 
   self.animatorOut = function(elem, isInit) {
-    if (!isInit) {
-      elem.onclick = function(ev) {
-        ev.preventDefault();
-        vhxm.components.shared.sidebar.state.isOpen(false);
-
-        $('.c-sidebar').velocity({
-          right: '-470'
-        }, {
-          duration: 500,
-          easing: [ 0.19, 1, 0.22, 1 ],
-          complete: function() {
-            if (vhxm.components.shared.sidebar.state.animate_out_cb) {
-              vhxm.components.shared.sidebar.state.onAnimateOut();
-            }
-          }
-        });
+    $(elem).velocity({
+      right: '-470'
+    }, {
+      duration: 500,
+      easing: [0.19, 1, 0.22, 1],
+      complete: function complete() {
+        vhxm.components.shared.sidebar.state.onClose();
+        vhxm.components.shared.sidebar.state.skipTransition(false);
       }
-    }
+    });
   };
 };