+++ quartz-js/components/sidebar/sidebar.state.js
@@ -1,8 +1,8 @@
 vhxm.components.shared.sidebar.state = {
   isOpen         : m.prop(false),
+  isLoaded       : m.prop(false),
   skipTransition : m.prop(false),
-  model          : m.prop(null),
   template       : m.prop(null),
-  onAnimateOut   : m.prop(null),
-  onAnimateIn    : m.prop(null)
+  onClose        : m.prop(function() {}),
+  onOpen         : m.prop(function() {})
 };